@@ -1,13 +1,13 @@
 import 'package:firexcode/firexcode.dart';
 
 class TextView extends StatefulWidget {
-  final double left;
-  final double top;
-  final Function ontap;
-  final Function(DragUpdateDetails) onpanupdate;
-  final Map mapJson;
+  final double? left;
+  final double? top;
+  final Function? ontap;
+  final Function(DragUpdateDetails)? onpanupdate;
+  final Map? mapJson;
   const TextView({
-    Key key,
+    Key? key,
     this.left,
     this.top,
     this.ontap,
@@ -21,21 +21,21 @@ class TextView extends StatefulWidget {
 class _TextViewState extends State<TextView> {
   @override
   Widget build(BuildContext context) {
-    return widget.mapJson['name']
+    return widget.mapJson!['name']
         .toString()
         .text(
-            textAlign: widget.mapJson['align'],
+            textAlign: widget.mapJson!['align'],
             style: TextStyle(
-              color: widget.mapJson['color'],
-              fontSize: widget.mapJson['size'],
+              color: widget.mapJson!['color'],
+              fontSize: widget.mapJson!['size'],
             ))
         .xGesture(
-          onTap: widget.ontap,
-          onPanUpdate: widget.onpanupdate,
+          onTap:()=> widget.ontap,
+          onPanUpdate: widget.onpanupdate!,
         )
         .xPositioned(
-          left: widget.left,
-          top: widget.top,
+          left: widget.left!,
+          top: widget.top!,
         );
   }
 }
