@@ -3,9 +3,9 @@ import 'dart:io';
 import 'package:firexcode/firexcode.dart';
 
 class ImageView extends StatefulWidget {
-  final File file;
+  final File? file;
 
-  const ImageView({Key key, this.file}) : super(key: key);
+  const ImageView({Key? key, this.file}) : super(key: key);
   @override
   _ImageViewState createState() => _ImageViewState();
 }
@@ -13,6 +13,6 @@ class ImageView extends StatefulWidget {
 class _ImageViewState extends State<ImageView> {
   @override
   Widget build(BuildContext context) {
-    return Image.file(widget.file).xCenter().xScaffold();
+    return Image.file(widget.file!).xCenter().xScaffold();
   }
 }
